requirements.txt
flask==2.3.2
tensorflow==2.20.0
numpy==1.24.3
pillow==9.5.0
opencv-python==4.8.0.74
gunicorn==21.2.0
